Remove `CheckBalanceJob` (#11295)
## Summary of the problem

-
https://appsignal.com/hack-club/sites/6596247683eb67648f30f807/exceptions/incidents/2026
- https://hackclub.slack.com/archives/C047Y01MHJQ/p1754931278490929

This job is creating a lot of error reporting noise. It used to be a lot
more useful when we used Airbrake which we could configure to send us
daily reports, but at the moment it's not adding any value beyond what
our current balance tracking mechanisms provide.

## Describe your changes

Delete the job and the code that triggers it.