[Categories] Set automatically based on stripe merchant codes (#11386)
## Summary of the problem

In #11298 and #11371 we add a mechanism to associate transactions to
categories. We would like to do that automatically for Stripe
transactions.

## Describe your changes

1. Adds a `merchant_category` method to both the raw stripe transaction
models
2. Adds `TransactionCategoryService#sync_from_stripe!` to handle the
syncing logic
3. Runs (2) when `CanonicalPendingTransaction`s are created for
`RawPendingStripeTransaction`s

I _think_ this is sufficient given the updated logic in
`CanonicalPendingTransactionService::Settle` but I'm not sure if that's
the only way Stripe transactions are created.