[Announcement] Cancel monthly announcements with no data (#11257)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
Right now, monthly announcements are enabled for a lot of transparent
organizations, including some that may not have any activity. We don't
want to send out all these monthly announcements with empty data, like
if the organization has no donations.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Any announcements that are still template drafts at the end of the month
will not be published. At the 7 day mark, monthly announcements that
have data in all of their blocks will be promoted to drafts, which will
allow them to be sent. All other monthly announcements will be left as
template drafts and a notification will be sent out to managers that if
they do not edit the announcement, it will not be sent out. This is also
displayed in a callout on the announcements page.

<img width="933" height="352" alt="image"
src="https://github.com/user-attachments/assets/7aa9c6c3-5b77-4efd-b148-17bf7c81779b"
/>


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->